Hello World
To write "Hello, World!" in Java, you can create a simple Java class with a main method.

Here's a breakdown of the code:

public class HelloWorld: This declares a class named HelloWorld. In Java, the name of the class must match the filename.

public static void main(String[] args): This is the entry point of the program. It's a special method that gets executed when you run the program. The String[] args is an array of command-line arguments (which you can ignore for this simple example).

System.out.println("Hello, World!");: This line prints the "Hello, World!" message to the console. System.out refers to the standard output stream, and println is a method that prints a line of text.

Save the above code in a file named HelloWorld.java. Then, you can compile and run it using the following commands in the terminal or command prompt: