Types of Loop
In Java, there are three main types of loops: for, while, and do-while. Each loop type serves a different purpose and can be used in various situations.
1. for Loop:
The for loop is used when the number of iterations is known beforehand. It has a more compact syntax and is commonly used for iterating over arrays or a range of values.

2. while Loop:
The while loop is used when the number of iterations is not known beforehand, and it depends on a condition. The loop continues executing as long as the specified condition is true.

3. do-while Loop:
The do-while loop is similar to the while loop, but it guarantees that the code inside the loop will be executed at least once, even if the condition is initially false.

And their is one but this is combination of Among Above anyone

4. Nested Loop
A nested loop in programming refers to placing one loop inside another. This allows you to create more complex and structured iterations. Nested loops are often used when you need to perform repetitive tasks within another repetitive task