Conditional Statements
In Java, conditional statements are used to control the flow of the program based on certain conditions. The most common conditional statements are if, else if, and else. Here's a brief overview of how they work:

1. if Statement:
The if statement is used to execute a block of code only if a specified condition is true.
2. if-else Statement:
The if-else statement allows you to execute one block of code if the condition is true and another block if the condition is false.
3. if-else if-else Statement:
The if-else if-else statement allows you to check multiple conditions and execute the block of code corresponding to the first true condition.
4. Nested if Statements:
You can also have if statements inside other if statements, creating nested conditions.
5. Switch Statement:
The switch statement is used when you have multiple possible conditions to check against a single variable.

These conditional statements allow you to control the flow of your Java program based on different conditions.