Arrays
An array in Java is a data structure that allows you to store multiple values of the same type in a single variable. It is a container object that holds a fixed number of elements of the same data type. Each element in an array is accessed by its index, starting from 0 for the first element. The size of an array is fixed at the time of its creation, and it cannot be changed later.

2D Arrays in Java:

A 2D array in Java is an array of arrays. It can be thought of as a grid or matrix where each element is identified by two indices – the row and the column. In Java, you can declare, create, and initialize a 2D array using the following syntax: